backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
httpx==0.28.1
pydantic-settings==2.7.1
psycopg2-binary==2.9.10
sqlparse==0.5.3
pytest==8.3.5
pytest-asyncio==0.25.3
